machine-learning-main/requirements.txt
# Use versions that provide pre-compiled wheels for Python 3.13
streamlit>=1.35.0
firebase-admin>=6.5.0
google-generativeai>=0.5.0
spacy>=3.7.0
groq>=0.5.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
duckduckgo-search>=5.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.1
nltk>=3.8.0
requests>=2.31.0

# FastAPI + Celery for backend microservices
fastapi>=0.104.0
uvicorn>=0.24.0
celery>=5.3.0
redis>=5.0.0
websockets>=11.0.0
neo4j
